unit_with_power.rs private module
