+++ src-tauri/src/units/unit.rs
@@ -1,14 +1,13 @@
 
 use super::dimension::Dimension;
+use super::unit_with_power::UnitWithPower;
 use super::composite::CompositeUnit;
 
-use num::One;
 use num::pow::Pow;
 use thiserror::Error;
 
 use std::fmt::{self, Formatter, Display};
 use std::ops::{Mul, Div};
-use std::cmp::Ordering;
 
 /// A unit is a named quantity in some [`Dimension`] which can be
 /// converted to the "base" unit of that dimension.
@@ -37,13 +36,6 @@ pub struct Unit<T> {
   /// consist only of simple units and have the same dimension as
   /// `self`. These preconditions are enforced by this API.
   composed_units: Option<Box<CompositeUnit<T>>>,
-}
-
-/// A named unit raised to an integer power.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct UnitWithPower<T> {
-  pub unit: Unit<T>,
-  pub exponent: i64,
 }
 
 /// Error type returned from [`Unit::with_composition`].
@@ -175,75 +167,9 @@ impl<T> Unit<T> {
   }
 }
 
-impl<T> UnitWithPower<T> {
-  pub fn dimension(&self) -> Dimension {
-    self.unit.dimension().pow(self.exponent)
-  }
-
-  pub fn to_base<'a, U>(&'a self, mut amount: U) -> U
-  where U: Mul<&'a T, Output = U>,
-        U: Div<&'a T, Output = U> {
-    match self.exponent.cmp(&0) {
-      Ordering::Greater => {
-        for _ in 0..self.exponent {
-          amount = self.unit.to_base(amount);
-        }
-      }
-      Ordering::Less => {
-        for _ in 0..(-self.exponent) {
-          amount = self.unit.from_base(amount);
-        }
-      }
-      Ordering::Equal => {},
-    }
-    amount
-  }
-
-  pub fn from_base<'a, U>(&'a self, mut amount: U) -> U
-  where U: Mul<&'a T, Output = U>,
-        U: Div<&'a T, Output = U> {
-    match self.exponent.cmp(&0) {
-      Ordering::Greater => {
-        for _ in 0..self.exponent {
-          amount = self.unit.from_base(amount);
-        }
-      }
-      Ordering::Less => {
-        for _ in 0..(-self.exponent) {
-          amount = self.unit.to_base(amount);
-        }
-      }
-      Ordering::Equal => {},
-    }
-    amount
-  }
-}
-
-impl<T> From<Unit<T>> for CompositeUnit<T> {
-  fn from(unit: Unit<T>) -> Self {
-    CompositeUnit::new([UnitWithPower { unit, exponent: 1 }])
-  }
-}
-
-impl<T> From<UnitWithPower<T>> for CompositeUnit<T> {
-  fn from(unit: UnitWithPower<T>) -> Self {
-    CompositeUnit::new([unit])
-  }
-}
-
 impl<T> Display for Unit<T> {
   fn fmt(&self, f: &mut Formatter) -> fmt::Result {
     write!(f, "{}", self.name)
-  }
-}
-
-impl<T> Display for UnitWithPower<T> {
-  fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-    if self.exponent == 1 {
-      write!(f, "{}", self.unit)
-    } else {
-      write!(f, "{}^{}", self.unit, self.exponent)
-    }
   }
 }
 
@@ -273,36 +199,10 @@ where S: Into<CompositeUnit<T>> {
   }
 }
 
-impl<T, S> Mul<S> for UnitWithPower<T>
-where S: Into<CompositeUnit<T>> {
-  type Output = CompositeUnit<T>;
-
-  fn mul(self, rhs: S) -> Self::Output {
-    CompositeUnit::from(self) * rhs
-  }
-}
-
-impl<T, S> Div<S> for UnitWithPower<T>
-where S: Into<CompositeUnit<T>> {
-  type Output = CompositeUnit<T>;
-
-  fn div(self, rhs: S) -> Self::Output {
-    CompositeUnit::from(self) / rhs
-  }
-}
-
-impl<T> Pow<i64> for UnitWithPower<T> {
-  type Output = UnitWithPower<T>;
-
-  fn pow(self, rhs: i64) -> Self::Output {
-    UnitWithPower { unit: self.unit, exponent: self.exponent * rhs }
-  }
-}
-
 #[cfg(test)]
 mod tests {
   use super::*;
-  use crate::units::dimension::{Dimension, BaseDimension};
+  use crate::units::dimension::BaseDimension;
   use crate::units::test_utils::{meters, kilometers, seconds, minutes};
 
   #[test]
@@ -356,35 +256,5 @@ mod tests {
       .try_with_composed(CompositeUnit::new(vec![UnitWithPower { unit: kilometers(), exponent: 2 }]))
       .unwrap_err();
     assert_eq!(err.reason, UnitCompositionErrorReason::DimensionMismatch);
-  }
-
-  #[test]
-  fn test_unit_with_power_dimension() {
-    let unit = UnitWithPower { unit: kilometers(), exponent: 3 };
-    assert_eq!(unit.dimension(), Dimension::singleton(BaseDimension::Length).pow(3));
-    let unit = UnitWithPower { unit: seconds(), exponent: -2 };
-    assert_eq!(unit.dimension(), Dimension::singleton(BaseDimension::Time).pow(-2));
-    let unit = UnitWithPower { unit: kilometers(), exponent: 0 };
-    assert_eq!(unit.dimension(), Dimension::one());
-  }
-
-  #[test]
-  fn test_unit_with_power_to_base() {
-    let unit = UnitWithPower { unit: kilometers(), exponent: 3 };
-    assert_eq!(unit.to_base(2.0), 2_000_000_000.0);
-    let unit = UnitWithPower { unit: kilometers(), exponent: -1 };
-    assert_eq!(unit.to_base(2_000.0), 2.0);
-    let unit = UnitWithPower { unit: kilometers(), exponent: 0 };
-    assert_eq!(unit.to_base(199.0), 199.0);
-  }
-
-  #[test]
-  fn test_unit_with_power_from_base() {
-    let unit = UnitWithPower { unit: kilometers(), exponent: 3 };
-    assert_eq!(unit.from_base(2_000_000_000.0), 2.0);
-    let unit = UnitWithPower { unit: kilometers(), exponent: -1 };
-    assert_eq!(unit.from_base(2.0), 2_000.0);
-    let unit = UnitWithPower { unit: kilometers(), exponent: 0 };
-    assert_eq!(unit.from_base(199.0), 199.0);
   }
 }
